requirements.txt
certifi==2018.1.18
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11
django-admin-rangefilter==0.3.7
django-appconf==1.0.2
django-bootstrap3==9.1.0
django-cleanup==2.1.0
django-cors-headers==2.2.0
django-filter==1.1.0
django-heroku==0.3.1
django-imagekit==4.0.2
django-mailjet==0.3.1
django-material==1.1.0
django-taggit==0.23.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
drf-generators==0.3.0
future==0.10.1
gunicorn==19.9.0
idna==2.6
mailjet==1.4.1
mailjet-rest==1.3.0
MySQL-python==1.2.5
numpy==1.16.2
openpay==0.3
pilkit==2.0
Pillow==5.0.0
postgres==2.2.2
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
PyJWT==1.6.4
pytz==2018.3
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
whitenoise==4.1.1
